self-health-core-backend/user-service/src/main/kotlin/com/self/health/user/security/config/SecurityConfiguration.kt
package com.self.health.user.security.config

import org.springframework.context.annotation.Bean
import org.springframework.context.annotation.Configuration
import org.springframework.http.HttpStatus.UNAUTHORIZED
import org.springframework.security.config.annotation.web.builders.HttpSecurity
import org.springframework.security.config.http.SessionCreationPolicy
import org.springframework.security.web.SecurityFilterChain
import org.springframework.security.web.authentication.HttpStatusEntryPoint


@Configuration
class SecurityConfiguration {
    @Bean
     fun filterChain(http: HttpSecurity): SecurityFilterChain
        {
        return http.cors().and()
            .csrf().disable()
            .authorizeHttpRequests().anyRequest().permitAll()
            .and()
            .sessionManagement()
            .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
            .and()
            .exceptionHandling()
            .authenticationEntryPoint(HttpStatusEntryPoint (UNAUTHORIZED))
            .and().build()
    }
}
